requirements.txt
```
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfminer.six>=20211012
pandas>=2.0.0
python-dateutil>=2.8.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytz>=2023.3
lxml>=4.9.0
```